pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.19.0",
    "aiosqlite>=0.21.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
